Further refactoring of SMVC's System.
