+++ system/Helpers/Database.php
@@ -10,7 +10,7 @@
 
 namespace Smvc\Helpers;
 
-use Smvc\Core\Config;
+use Smvc\Config;
 
 use PDO;
